separate BlendShapeBinding and MaterialValueBinding
