+++ Assets/UniVRM-1.0/Components/BlendShape/BlendShapeBinding.cs
@@ -1,0 +1,39 @@
+using System;
+
+namespace UniVRM10
+{
+    [Serializable]
+    public struct BlendShapeBinding : IEquatable<BlendShapeBinding>
+    {
+        public String RelativePath;
+        public int Index;
+        public float Weight;
+
+        public override string ToString()
+        {
+            return string.Format("{0}[{1}]=>{2}", RelativePath, Index, Weight);
+        }
+
+        public bool Equals(BlendShapeBinding other)
+        {
+            return string.Equals(RelativePath, other.RelativePath) && Index == other.Index && Weight.Equals(other.Weight);
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj)) return false;
+            return obj is BlendShapeBinding && Equals((BlendShapeBinding)obj);
+        }
+
+        public override int GetHashCode()
+        {
+            unchecked
+            {
+                var hashCode = (RelativePath != null ? RelativePath.GetHashCode() : 0);
+                hashCode = (hashCode * 397) ^ Index;
+                hashCode = (hashCode * 397) ^ Weight.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}
