+++ Assets/UniVRM-1.0/Components/BlendShape/BlendShapeBinding.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a9718a1a32a0ce4a9deaead86dbf393
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
