+++ Assets/UniVRM-1.0/Components/BlendShape/MaterialValueBinding.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 985d3a33d6e0e144e99deed88e048b59
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
